finished plugin

+++ src/me/becja10/PvPModerator/BlockedPlayer.java
@@ -9,13 +9,13 @@ public class BlockedPlayer {
 	
 	public static final int NEW_PLAYER = 0;
 	public static final int TP_EVENT = 1;
-	
+		
 	public BlockedPlayer(long l, int r){
 		time = l;
 		reason = r;
 	}
 	
-	public String getWarnMessage(boolean isTarget)
+	public String getWarnMessage(boolean isTarget, BlockedPlayer vic)
 	{
 		String ret = "";
 		
@@ -31,13 +31,19 @@ public class BlockedPlayer {
 			}
 		}
 		else{
-			switch(reason){
-			case NEW_PLAYER:
-				ret = ChatColor.RED + "This player is new! You can't kill them yet.";
-				break;
-			case TP_EVENT:
-				ret = ChatColor.RED + "This player is in tp cooldown. You can't attack them yet.";
-				break;
+			if(vic == null){
+				ret = ChatColor.RED + "You cannot PvP while protected. Use " + ChatColor.YELLOW + 
+						"/removeprotection" + ChatColor.RED + " if you wish to PvP";
+			}
+			else{			
+				switch(reason){
+				case NEW_PLAYER:
+					ret = ChatColor.RED + "This player is new! You can't kill them yet.";
+					break;
+				case TP_EVENT:
+					ret = ChatColor.RED + "This player is in tp cooldown. You can't attack them yet.";
+					break;
+				}
 			}
 		}
